Remove root logger handler
Since monocle is used as a library, we should let the calling
application  define the log handlers. logging.basicConfig adds a
handler to the root logger, which can cause unexpected duplicate logs
to appear in calling application. Remove it, and let the calling
application define handlers for monocle if necessary.